Relocate counter code into it's own file.
